Add from_secret and from_secret_ref functions to struct directly
