+++ stackable-certs/Cargo.toml
@@ -9,7 +9,6 @@ repository.workspace = true
 [features]
 default = []
 rustls = ["dep:tokio-rustls", "dep:rustls-pemfile"]
-k8s = ["dep:kube", "dep:k8s-openapi"]
 
 [dependencies]
 stackable-operator = { path = ".." }
@@ -17,10 +16,10 @@ stackable-operator = { path = ".." }
 const-oid = "0.9.6"
 ecdsa = { version = "0.16.9", features = ["digest", "pem"] }
 p256 = { version = "0.13.2", features = ["ecdsa"] }
-k8s-openapi = { version = "0.21.0", optional = true, default-features = false, features = [
+k8s-openapi = { version = "0.21.0", default-features = false, features = [
   "v1_28",
 ] }
-kube = { version = "0.88.1", optional = true, default-features = false, features = [
+kube = { version = "0.88.1", default-features = false, features = [
   "client",
   "rustls-tls",
 ] }
